backend/pyproject.toml
[project]
name = "darcy-hackathon-backend"
version = "0.1.0"
description = "Darcy Hackathon — FastAPI backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic-settings>=2.7.0",
    "anthropic>=0.52.0",
    "sse-starlette>=2.0.0",
    "qdrant-client>=1.12.0",
    "openai>=1.60.0",
    "rank-bm25>=0.2.2",
    "aiosqlite>=0.20.0",
    "python-multipart>=0.0.18",
    "pypdf>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.28.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
]

[tool.ruff]
target-version = "py312"
line-length = 88
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]

[tool.ruff.lint.per-file-ignores]
"src/services/claude_agent.py" = ["E501"]
"src/services/consultant_agent.py" = ["E501"]
"src/routes/*.py" = ["B008"]  # FastAPI uses Depends/Query in default args

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = "rank_bm25"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pypdf"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
